Add WalletFactory
WalletFactory is a contract which creates wallets for a given client. The goal is to be a simple contract that can be called when we need to create a new wallet. For any changes to the default wallet or addresses, we will simply create a new wallet factory and point our new registration calls to this point. The goal is that our clients own the wallets we create, but we may establish a small amount of rights to, for instance, allow auto-pay on loans.