Add `NewWallet` Event and change wallets to be owned by `msg.sender`
This patch changes wallet factory to add a new `NewWallet` event when new wallets are created. Additionally, we now expect wallets to be created by the `msg.sender`. Both of these changes are in line with how we expect to create wallets in the new user registration flow. For more details, follow this issue: https://github.com/compound-finance/compound/pull/21