We own WalletFactory so that we will pay gas for Wallet creation
As the user account will not have gas at the time wallets are created, we will bear the burden of creating the wallet for the user. As such, we change our functions to allow us to create a wallet on behalf of a user. We will fund the user account with a gas stipend only after the user has initiated a deposit into Compound.