Add address(self) to new wallet event
This patches adds `address(self)` as the third argument to the NewWallet event. This is because we want to know exactly which wallet was created for a user, and 1) during transition periods it may change from what we expect and what was created, and 2) the address is not currently passed to our Farmer handlers (though #1 is a better argument overall). Thus, we opt to add this to the event.