Deploy `EtherToken`, `WalletFactory` and `Bank` contracts on deploy
This patch updates our deploy code to make sure we deploy all three of the above contracts to the blockchain. We may opt in the future to use an existing `EtherToken` contract, but that's a long term plan. Note: `WalletFactory` depends on both `EtherToken` and `Bank` and thus is deployed last.