[WIP] Change to a ledger entry model
This patch starts to move our entire system to a ledger entry model. This approach allows us to be very consistent with our accounting events using the correct real-world terms. Additonally, we start to make loans nothing more than a variety of account with opposite effects as a savings account (e.g. crediting a loan decreases the balance).