Add depositDirect Function to move assets from wallet to Vault
This patch adds a `depositDirect` function which moves assets from the Wallet into the Vault. This is for when a customer moves his/her tokens to the Wallet. We detect that the user moved these tokens (via watching the ERC-20 contract) if (at our discretion) call this function to move these tokens into the Vault (where they belong!). Note: this function is public since tokens should never sit into the Wallet (and this function *only* lets you move them into the Vault).