Add ErrorCodes instead of Throwing in most cases (#50)
* Add ErrorCodes instead of Throwing in most cases

This patch adds a new concept (Graceful) which is focused around us emitting error codes via return values (0=success, else=failure) instead of throwing when checks and conditions fail. This requires a bit more brainpower to make sure no effects have occured, but by and large, this means we actually know how and why a contract failed. This is a powerful concept, but I am open to suggestions in how to make this better.

* Start to re-work to use boolean return values and graceful failure logs

* Improve the rest of graceful failures, get nearly everything else passing.

* continue to fix test cases

* Continue to fix tests for better GracefulFailures

* Change from return failure() to failure() and return false

* Add back missing await

* Fix utils

* Spacing fixes

* Use promises instead of async/await

* Fix spelling

* Remove unused enums