Start to build storage contracts
This patchs starts to build out contracts which are used for long-term storage. The goal is to keep the logic as light-weight as possible in these contracts since they will not be upgradable. This allows us to upgrade non-storage contracts and point them at the existing storage contracts. We may decide to allow or not allow upgrades for production builds, but this is a strict requirement for us to push changes in a live staging environment.