Start to build storage contracts (#55)
* Start to build storage contracts

This patchs starts to build out contracts which are used for long-term storage. The goal is to keep the logic as light-weight as possible in these contracts since they will not be upgradable. This allows us to upgrade non-storage contracts and point them at the existing storage contracts. We may decide to allow or not allow upgrades for production builds, but this is a strict requirement for us to push changes in a live staging environment.

* Add TokenStore to long-term hold tokens since Vault can be redeployed

* start to add a full suite of unit tests for storage contracts

* Fix the rest of test cases and get migrations working that do not overwrite storage

* Add a fully sussed out deployment scheme

* try to persist workspaces a bit more by hand

* Nice quites

* Don't exit 0

* Also use deployed builds for follow-up steps

* Use caching to keep consistent values

* Es la llave

* Don't use full full cache key

* Start over

* Add back in cached deploys

* Just do everything as callbacks

* Add back in test cases

* Fix Mission network address

* Re-arrange how these are deployed

* Temporarily do clean builds

* Fix perplexing issue of Truffle choosing mission net instead of test net

* Convert indention to tabs

* Minor updates to pass tests