Add Interest Rates via Daily Computation (#56)
* Add Interest Rates via Daily Computation

This patch adds a new Snapshot mechanism to the InterestRateStorage contract. We now track the interest rate each day and what the compound interest would be from that given day to check up to the current day (e.g. principal * getCompoundedInterest(depositTimestamp) = current balance). We do this by storing each day's interest rate (as of midnight) in a mapping from "day # -> Snapshot". Here, day # is the number of days since January 1, 1970 (which is unix timestamp / `60*60*24` truncated). This makes it trivial to find a given snapshot. Each snapshot stores the current timestamp (when it was snapshotted), the current daily interest rate (scaled by 10e9) and the compounded interest rate (that number above). Each time we make a new snapshot, we walk all previous snapshots and update that Snapshot by multiplying its compounded interest rate by the current compounded interest rate to give us the correct multiplier from that day. We keep each number scaled by 10e9 such that this number will not grow unbounded.

* Switch to Block-based Interest and make sure to fill in any missing block units

* Interest Rate Formula and Storage / Balance Sheet (#58)

* Start to fix test cases for upgraded balance sheet storage

* Continue to build out nearly fully functional interest rate functions

* Continue to fix items for Vault tests

* Fix test cases for new Interest Rate Storage method

* Add proper addresses.json