Update Interest Rate Model
This patch updates the interest rate model to use the formula specified in the current version of the white paper (essentially replacing 1-`reserve ratio` with simply `borrows/supply`). To prevent the rates from asymptoting, we'll need to add cash balance checks on new borrows, but otherwise, this new formula should be a drop-in replacement for our current system.