Start to rebuild BalanceSheet as new contract
This patch starts to rebuild the balance sheet as a new contract itself. This will improve our ability to track Supply, Debits and Cash. This is frustrating because the design limitations of Solidity constantly get in the way of writing code that's as coherent and testable as I'd like, but, so it goes. Currently working on getting MoneyMarket.js and Wallet.js to pass tests.